Update patch set 4
Patch Set 4:

I don't think we need to show the version of the app in title as there is already one in the new tab.Sorry for my poor English.
But I described it in the commit message, maybe we can leave the code as a comment or simply ignore it.

Patch-set: 4